Move ExceptionHandlerTrait under Node namespace
